Rename platform specific interfaces
While these are approximately 1:1 with the different platforms, it is
more appropriate to define them by their interface name so that they can
be reused between, for example, multiple platforms that use X11.
