+++ lib/opengl_es.rb
@@ -3,7 +3,7 @@ require_relative 'opengl_common'
 require_relative 'opengl_es_enum'
 require_relative 'opengl_es_command'
 if GL.get_platform == :OPENGL_PLATFORM_WINDOWS
-  require_relative 'opengl_windows'
+  require_relative 'opengl_interface_wgl'
 end
 
 =begin
