+++ lib/opengl_interface_cgl.rb
