+++ lib/opengl_interface_glx.rb
