+++ lib/opengl_interface_wgl.rb
